+++ encryption_key_store/api/upload/router.py
@@ -16,38 +16,36 @@
 
 from fastapi import APIRouter, status
 
-from . import upload_exceptions, upload_models
+from . import exceptions, models
 from .envelope_decryption import extract_envelope_content, store_secret
 
-router = APIRouter()
+upload_router = APIRouter()
 
 ERROR_RESPONSES = {
     "malformedOrMissingEnvelope": {
         "description": (""),
-        "model": upload_exceptions.HttpMalformedOrMissingEnvelopeError.get_body_model(),
+        "model": exceptions.HttpMalformedOrMissingEnvelopeError.get_body_model(),
     },
     "envelopeDecryptionError": {
         "description": (""),
-        "model": upload_exceptions.HttpEnvelopeDecrpytionError.get_body_model(),
+        "model": exceptions.HttpEnvelopeDecrpytionError.get_body_model(),
     },
 }
 
 
-@router.post(
+@upload_router.post(
     "/envelopes",
     summary="Extract file encryption/decryption secret and file content offset from enevelope",
     operation_id="postEncryptionData",
     status_code=status.HTTP_200_OK,
-    response_model=upload_models.InboundEnvelopeContent,
+    response_model=models.InboundEnvelopeContent,
     response_description="",
     responses={
         status.HTTP_400_BAD_REQUEST: ERROR_RESPONSES["malformedOrMissingEnvelope"],
         status.HTTP_403_FORBIDDEN: ERROR_RESPONSES["envelopeDecryptionError"],
     },
 )
-async def post_encryption_secrets(
-    *, envelope_query: upload_models.InboundEnvelopeQuery
-):
+async def post_encryption_secrets(*, envelope_query: models.InboundEnvelopeQuery):
     """Extract file encryption/decryption secret, create secret ID and extract
     file content offset"""
     user_id = envelope_query.user_id
@@ -56,10 +54,8 @@ async def post_encryption_secrets(
     except ValueError as error:
         # Everything in crypt4gh is an ValueError... try to distinguish based on message
         if "No supported encryption method" == str(error):
-            raise upload_exceptions.HttpEnvelopeDecrpytionError() from error
-        raise upload_exceptions.HttpMalformedOrMissingEnvelopeError(
-            user_id=user_id
-        ) from error
+            raise exceptions.HttpEnvelopeDecrpytionError() from error
+        raise exceptions.HttpMalformedOrMissingEnvelopeError(user_id=user_id) from error
 
     secret_id = await store_secret(secret)
     return {"secret": secret, "secret_id": secret_id, "offset": offset}
